clean and structure python code, add image plot alongside recognition
