+++ py_src/__init__.py
@@ -1,0 +1,4 @@
+from .params import COCO_INSTANCE_CATEGORY_NAMES
+from .utils import download_image, load_tensor_and_image
+from .ai_callable.pretrained import detect
+from .ai_callable.train import train
