+++ py_src/ai_callable/train.py
@@ -1,0 +1,2 @@
+def train(args, device):
+    pass
