+++ py_src/params/__init__.py
@@ -1,0 +1,21 @@
+# COCO_INSTANCE_CATEGORY_NAMES = [
+#     '__background__', 'person', 'bicycle', 'car', 'motorcycle', 'airplane',
+#     'bus', 'train', 'truck', 'boat', 'traffic light', 'fire hydrant', 'N/A',
+#     'stop sign', 'parking meter', 'bench', 'bird', 'cat', 'dog', 'horse',
+#     'sheep', 'cow', 'elephant', 'bear', 'zebra', 'giraffe', 'N/A', 'backpack',
+#     'umbrella', 'N/A', 'N/A', 'handbag', 'tie', 'suitcase', 'frisbee', 'skis',
+#     'snowboard', 'sports ball', 'kite', 'baseball bat', 'baseball glove',
+#     'skateboard', 'surfboard', 'tennis racket', 'bottle', 'N/A', 'wine glass',
+#     'cup', 'fork', 'knife', 'spoon', 'bowl', 'banana', 'apple', 'sandwich',
+#     'orange', 'broccoli', 'carrot', 'hot dog', 'pizza', 'donut', 'cake',
+#     'chair', 'couch', 'potted plant', 'bed', 'N/A', 'dining table', 'N/A',
+#     'N/A', 'toilet', 'N/A', 'tv', 'laptop', 'mouse', 'remote', 'keyboard',
+#     'cell phone', 'microwave', 'oven', 'toaster', 'sink', 'refrigerator',
+#     'N/A', 'book', 'clock', 'vase', 'scissors', 'teddy bear', 'hair drier',
+#     'toothbrush'
+# ]
+COCO_INSTANCE_CATEGORY_NAMES = ['__background__', 'aeroplane', 'bicycle',
+                                'bird', 'boat', 'bottle', 'bus', 'car', 'cat',
+                                'chair', 'cow', 'diningtable', 'dog', 'horse',
+                                'motorbike', 'person', 'pottedplant',
+                                'sheep', 'sofa', 'train', 'tvmonitor']
